Quick view UI: Fix to support multiple quick links
There was previously a limitation in the detail view/quickview where
only 1 'view all' link was visible per page. Change this to show n
number of view all links, if an array is passed to 'viewAll.' This also
lowers the # of clicks necessary to navigate to some items.

This also adds more view all buttons to instances, storage, and system
resources.