Fix for _pingTimeout being 0 in AgentMonitor
With commit d79f1f6fdc8307aa4038bfb2c7607904b89eedbe the AgentMonitor
was replaced with a pluggable service. However the ping timeout in the
original constructor was not passed on anymore, leading to a default
pingTimout of 0. This would fail all agents constantly.

Modified the startMonitor command to take a pingtimeout as an argument
and instruct AgentManagerImpl to pass it along.