Replace restart with SIGHUP
This signal will force the dnsmasq daemon to reload the configuration directly. This is much faster than restarting the daemon, which result in a much smaller window during which no dns server is available.

Tested by using the replaced version of edithosts.sh on a running vrouter causing dns problems.