Account object dereferences as account.name
Previously the jsonHelper would incorrectly deserialize the json
response if the object returned had an attribute of the same name
resulting in an additional level of indirection to refer to the entity's
attributes. eg: account.account.name instead of account.name

This fixes the tests that have still the same indirection exposed.

Signed-off-by: Prasanna Santhanam <tsp@apache.org>