Fix test_assign_vm to be discovered by nose testrunner
the nosetests runner will not discover tests unless the tests start with
the test_ prefix. When wrapping by the log_test_exceptions method the
name of the method changes and is not test_ anymore causing the runner
to skip the test. Included the __doc__ string as well so the runner can
report results correctly.

Signed-off-by: Prasanna Santhanam <tsp@apache.org>