CLOUDSTACK-728 Rewrite support for lswitch connected nics on both standard and distributed switches.
Fix the name of the interface id in extra config, so the nic is found up using the nic uuid instead of the vm id.

getVlanInfo should return null when the nic is plugged on a Lswitch network.

On a distributed portgroup all vms should be connected to a single
portgroup and set a specific vlan on the port. On a standard switch each nic should have its own portgroup with a dedicated vlan (not related to CS vlans)