CLOUDSTACK-3469 [VMware] Support to make username/password optional while adding VMware cluster
For legacy zones insist on the old model of asking for credentials for each cluster being added.
For non-legacy zones check if username & password are provided. If either or both not provided, try to retrieve & use the credentials from database, which are provided earlier while adding VMware DC to zone.
This lets user to not specify credentials while adding VMware cluster to zone, if the credentials are same as that of provided while adding VMware DC to zone.

Signed-off-by: Sateesh Chodapuneedi <sateesh@apache.org>