CLOUDSTACK-4310: Document vnc_listen for Qemu/libvirt
The vnc-listen patch didn't make it into 4.2 due to a newer version
of libvirt-java not being released.

Eventually we want to get rid of this, but currently we are forced to
do it this way. Setting it to anything else then 0.0.0.0 will break
migrations without a newer version of libvirt-java

The newer version of libvirt-java supports live migratons with a different
XML definition for on the destination host.