Add ability to programatically disable CloudStack ComponentConext intialization
As we moved to a more Spring managed lifecycle, the lifecycle management
in ComponentContext is not needed.  The transition to fully spring managed
will not happen all at once.  The management server will be done first and
then AWSAPI and UsageServer will come later.  This change allows the
ComponentContext in the context of the management server to be bypassed.
Note, by default, the ComponentContext will run its initialization.