CLOUDSTACK-5146: Removing basic and sg tags from the test  case which is invalid for the basic zone
TestRouterRestart.test_01_restart_network_cleanup test scenario is invalid for basic zone. The test case tries to restart the Router assigned to individual account but in basic zone, there is only one router i.e of shared network and there is only one guest(shared) network per zone. Hence the API fails to list the router of the individual account.

More over, we will not try to restart the Router of the shared network because it is used by many accounts and restarting it may affect the execution of other test cases.