Fix incorrect context being passed to hosts->instances section
Fixes issue where, if an action is being performed from the quick view in the
'Instances' section, the old context will be passed when viewing instances in
the 'hosts' section, causing all instances to be displayed. This was caused
by the quickview tooltip still being in the DOM, even after closing it by
clicking an action.