CLOUDSTACK-7467 Fix TestVolumes.test_07_resize_fail
Previously if you had a volume using a non customisable disk offering, and
attempted to resize it passing in the same disk offering id, the command would
be accepted, but it would actually be resized to its current size (i.e. the
provided size parameter was ignored). This is what the test used to check.

Commit 5df4ad6d0ad7e4233e03f2af1657639d07998746 modified the logic to check if
the provided diskofferingid was the same as the current one, and if so treat it
as if it hadn't been provided - this means the resize command now fails, which
is probably the more sensible thing to do (rather than giving the impression it
will be resized but actually not doing so).

This change therefore modifies the test logic to match.

Signed-off-by: Mike Tutkowski <mike.tutkowski@solidfire.com>