kvm: Add better logging when fetching a volume from libvirt
Clearly show if a volume is found and if not, that the pool is being refreshed
and the fetch is tried again.

Due to my commit be7ddb887ffbea17ca429dfc609ba3d618fa5c34 the chance of a volume
not being found is slightly bigger, but the performance gain is enormous on larger
deployments.

This is why we clearly have to log that we are refreshing the pool information
when a volume is not found.

It could be that a volume is created on host A and a few seconds later host B tries
to access the volume. In that case host B's libvirt doesn't know about the volume
yet and has to refresh the pool before it does.

(cherry picked from commit aca0523d3db0895b5e7a499ee782263dff92858a)