systemvmtemplate: Disable services that slow down boot
The console-setup service brings a nice font to the console, but why would we
want to use it. In most cases it takes a <10 seconds to set it up. When using
nested hypervising, I found this takes much longer time that causes tests to
time-out. I'd suggest turning off these services. They are not required for the
services the systemvm provides.

Manually picked from commit 5cc0acd
PR #254

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>