CLOUDSTACK-10016: VPC VR doesn't respond to DNS requests from remote access vpn clients (#2201)
When enabling remote access VPN, a new interface is created upon client connecting via VPN. The DNS service (dnsmasq) is set only to listen on interfaces that are active when it starts. Thus VPN users are provided the VR's IP address for DNS resolution, but it is not actually listening for DNS requests.